CHROMIUM: config: Enable CONFIG_PCIEPORTBUS config
The NVIDIA GPU uses platform-based power management to enter into
low power state and uses the virtual PME support. In the virtual
PME, whenever there is wake-up triggered from the host side, then
an eventfd will be generated and crosvm handles that. The crosvm
generates a virtual PME which will be handled by virtual root
port. To handle PME, the PCIEPORTBUS config needs to be enabled.

BUG=b:241526471
TEST=Run 'dmesg|grep OSC|grep PME' and check if PME is getting enabled

Signed-off-by: Abhishek Sahu <abhsahu@nvidia.com>
Signed-off-by: Vaibhav Vyas (SW) <vavyas@nvidia.com>
Change-Id: If29ee8236fe0615a67ec1edc49ff9cd79012e147
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4159798
Reviewed-by: Victor Ding <victording@chromium.org>