BACKPORT: mm: multi-gen LRU: simplify lru_gen_look_around()
Update the page generation in place with or without
current->reclaim_state->mm_walk.  The LRU lock is held for longer, if
mm_walk is NULL and the number of pages to update is more than
PAGEVEC_SIZE.

This causes a measurable regression from the LRU lock contention during a
microbencmark.  But a tiny regression is not worth the complexity.

Link: https://lkml.kernel.org/r/20230118001827.1040870-8-talumbau@google.com
Signed-off-by: T.J. Alumbaugh <talumbau@google.com>
Cc: Yu Zhao <yuzhao@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
(cherry picked from commit abf086721a2f1e6897c57796f7268df1b194c750)

BUG=b:266976439
TEST=built

Signed-off-by: Yu Zhao <yuzhao@google.com>
Change-Id: I01249636e6e64d08636cdca417af4132a2f080ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4392045
Tested-by: Yu Zhao <yuzhao@chromium.org>
Reviewed-by: T.J. Alumbaugh <talumbau@google.com>