FIXUP: CHROMIUM: LSM: chromiumos security module
0-day error:

>> include/linux/compiler_types.h:357:45: error: call to
   '__compiletime_assert_441' declared with attribute error:
   BUILD_BUG_ON failed: sizeof(*payload) != sizeof(prep->payload.data)

In case when 'CONFIG_SECURITY_CHROMIUMOS_NO_SYMLINK_MOUNT' is set to
y, the 'struct path' (in 'struct big_key_payload') gets an extra field
'int link_count', which the structure 'key_payload.data' (used as
prep->payload.data in function 'big_key_preparse') is not able to
fit in, because it does not expect 'struct path' to contain 'int
link_count'.

Increase the size of the array key_payload.data from 4 to 5 to make
up for the increased size of 'struct path' and instead of reporting bug
in the function 'big_key_preparse' when 'sizeof(*payload) !=
sizeof(prep->payload.data)', report bug when 'sizeof(*payload) >
sizeof(prep->payload.data)' to catch an error in case the size of
'struct path' increases in future.

BUG=chromium-os:21954,b:227610040
TEST=Use a mount path with a symlink and observe mount failure.

Reported-by: kernel test robot <lkp@intel.com>
Change-Id: I0ec38279c96bba0c3d3e4039acd058bddbd7fe8e
Signed-off-by: Aashish Sharma <shraash@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4222713
Reviewed-by: Guenter Roeck <groeck@chromium.org>