UPSTREAM: Bluetooth: hci_core: Detect if an ACL packet is in fact an ISO packet
Because some transports don't have a dedicated type for ISO packets
(see 14202eff214e1e941fefa0366d4c3bc4b1a0d500) they may use ACL type
when in fact they are ISO packets.

In the past this was left for the driver to detect such thing but it
creates a problem when using the likes of btproxy when used by a VM as
the host would not be aware of the connection the guest is doing it
won't be able to detect such behavior, so this make bt_recv_frame
detect when it happens as it is the common interface to all drivers
including guest VMs.

Fixes: 14202eff214e ("Bluetooth: btusb: Detect if an ACL packet is in fact an ISO packet")
Signed-off-by: Luiz Augusto von Dentz <luiz.von.dentz@intel.com>
(cherry picked from commit 876e78104f23ce9267822757a63562a609b126c3)
Signed-off-by: Jeremy Wu <jrwu@google.com>
Signed-off-by: Yun-hao Chung <howardchung@chromium.org>

BUG=b:314685767
TEST=build and test with LE audio local branch

Change-Id: I4c88104d7dad38ce756d60289647da3722d51e41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5320297
Commit-Queue: Yun-hao Chung <howardchung@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Yun-hao Chung <howardchung@chromium.org>
Reviewed-by: Ying Hsu <yinghsu@chromium.org>