CHROMIUM: i2c: designware: Fix clobbered i2c request status
In the newer Mendocino mbox implementation, the i2c request status read
from the PSP is getting clobbered. This causes bus arbitration issues as
the x86 will think the bus is free even though the PSP could be using
it.

Add a new variable to store the value of the mbox command register so
that the status of psp_smn_read does not clobber it.

Signed-off-by: Mark Hasemeyer<markhas@chromium.org>
Fixes: 23ffe4888017 (FROMLIST: i2c: designware: Add support for new SoCs
in AMDPSP driver - crrev.com/c/4004862)
Suggested-by: Sourabh Das<sourabh.das@amd.com>

BUG=b:259310959
UPSTREAM-TASK=b:274137251
TEST=tast run <dut-ip> hwsec.TPMContest

Change-Id: I8e5b59adf7d8de175946233239911d44b2fc260d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4348800
Tested-by: Mark Hasemeyer <markhas@google.com>
Commit-Queue: Mark Hasemeyer <markhas@google.com>
Reviewed-by: Tim Van Patten <timvp@google.com>
Reviewed-by: Jan Dabros <dabros@google.com>
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@google.com>