UPSTREAM: ASoC: qcom: q6prm: fix incorrect clk_root passed to ADSP
The second to last argument is clk_root (root of the clock), however the
code called q6prm_request_lpass_clock() with clk_attr instead
(copy-paste error).  This effectively was passing value of 1 as root
clock which worked on some of the SoCs (e.g. SM8450) but fails on
others, depending on the ADSP.  For example on SM8550 this "1" as root
clock is not accepted and results in errors coming from ADSP.

Fixes: 2f20640491ed ("ASoC: qdsp6: qdsp6: q6prm: handle clk disable correctly")
Cc: <stable@vger.kernel.org>
Signed-off-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Reviewed-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Tested-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Link: https://lore.kernel.org/r/20230302122908.221398-1-krzysztof.kozlowski@linaro.org
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 65882134bc622a1e57bd5928ac588855ea2e3ddd)

BUG=b:177043625
TEST=emerge-herobrine sys-kernel/chromeos-kernel-5_15

Signed-off-by: Linux Patches Robot <linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com>
Change-Id: Iefc84f4c795006e97c4257ee467cbdb1d16f7c40
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4352110
Tested-by: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Judy Hsiao <judyhsiao@google.com>