FROMGIT: pinctrl: amd: Disable and mask interrupts on resume
This fixes a similar problem to the one observed in:
commit 4e5a04be88fe ("pinctrl: amd: disable and mask interrupts on probe").

On some systems, during suspend/resume cycle firmware leaves
an interrupt enabled on a pin that is not used by the kernel.
This confuses the AMD pinctrl driver and causes spurious interrupts.

The driver already has logic to detect if a pin is used by the kernel.
Leverage it to re-initialize interrupt fields of a pin only if it's not
used by us.

Cc: stable@vger.kernel.org
Fixes: dbad75dd1f25 ("pinctrl: add AMD GPIO driver support.")
Signed-off-by: Kornel Dulęba <korneld@chromium.org>
Link: https://lore.kernel.org/r/20230320093259.845178-1-korneld@chromium.org
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
(cherry picked from commit b26cd9325be4c1fcd331b77f10acb627c560d4d7
 git.kernel.org/pub/scm/linux/kernel/git/linusw/linux-pinctrl fixes)

BUG=b:247634997
TEST=suspend_stress_test -c2 on zork

Change-Id: I63f50caebcdcfb6acc39ff00ba45cb56f3283c41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4384713
Tested-by: Kornel Dulęba <korneld@google.com>
Reviewed-by: Radoslaw Biernacki <biernacki@google.com>
Commit-Queue: Matt Davis <mattedavis@google.com>