FROMGIT: regulator: arizona-micsupp: Use PROBE_FORCE_SYNCHRONOUS
Restore synchronous probing for Arizona regulators as the main MFD
relies on the ordering of the devices probing.

As these regulators are built into the CODEC and typically have no DT
representation the regulator framework is unaware of their existence
until the driver probes. These means the probing of the driver needs to
be synchronous to ensure the regulators are not substitued for the dummy
later when the users request them.

Fixes: 259b93b21a9f ("regulator: Set PROBE_PREFER_ASYNCHRONOUS for drivers that existed in 4.14")
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Link: https://lore.kernel.org/r/20230323132047.833737-2-ckeepax@opensource.cirrus.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit b65a0a8edba2c0a03e8fdb03b6807132e4166483
 https://git.kernel.org/pub/scm/linux/kernel/git/broonie/regulator.git for-next)

BUG=b:272590035, b:167723050
TEST=Boot speed improves

Change-Id: Ide6b5e76f24cb6b4f106350b8f8cd24d9b0d2faf
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4367602
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>