UPSTREAM: media: venus: Warn only once about problems in .remove()
The only effect of returning an error code in a remove callback is that
the driver core emits a warning. The device is unbound anyhow.

As the remove callback already emits a (quite verbose) warning when ret
is non-zero, return zero to suppress the additional warning.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
(cherry picked from commit d80250819e59c76ee447fd58cbffb6f750a21918)

BUG=b:287490453
TEST=tast.video tests on Trogdor

Change-Id: I1858781ca23c87cea75b8f60457834d7657feb65
Signed-off-by: Nathan Hebert <nhebert@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4619669
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>