FROMGIT: drm/sysfs: Expose DRM connector id in each connector sysfs
Expose DRM connector id in device sysfs so that we can map the connector
id to the connector syspath. Currently, even if we can derive the
connector id from modeset, we do not have a way to find the
corresponding connector's syspath.

This is helpful when determining the root connector of MST tree. When a
tree of multiple MST hub is connected to the system, modeset describes
the tree in the PATH blob. For example, consider the following scenario.

+-------------+
| Source      |    +-------------+
| (Device)    |    | BranchX     |
|             |    | (MST)       |
|       [conn6]--->|       [port1]--->DisplayA
+-------------+    |             |
                   |             |    +-------------+
                   |             |    | BranchY     |
                   |             |    | (MST)       |
                   |       [port2]--->|       [port1]----->DisplayB
                   +-------------+    |             |
                                      |       [port2]----->DisplayC
                                      +-------------+

DPMST connector of DisplayA would have "mst:6-1" PATH.
DPMST connector of DisplayB would have "mst:6-2-1" PATH.
DPMST connector of DisplayC would have "mst:6-2-2" PATH.

Given that connector id of 6 is the root of the MST connector tree, we
can utilize this patch to parse through DRM connectors sysfs and find
which connector syspath corresponds to the root connector (id == 6).

ChromeOS intend to use this information for metrics collection. For
example, we want to tell which port is deriving which displays even with
a MST hub. Chromium patch for parsing DRM connector id from the kernel
is at http://crrev.com/c/4317207.

Signed-off-by: Won Chung <wonchung@google.com>
Reviewed-by: Manasi Navare <navaremanasi@chromium.org>
Signed-off-by: Manasi Navare <navaremanasi@chromium.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20230329014455.1990104-1-wonchung@google.com
(cherry picked from commit f0038cffa843697501c442eff4468e3ab37031f1
 git://anongit.freedesktop.org/drm/drm-misc drm-misc-next)

BUG=b:272149418
TEST=drm-misc-next branch

Change-Id: I3769e37a89df6f218305a4a0ccc0d03d19bfcde1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4497369
Tested-by: Won Chung <wonchung@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Won Chung <wonchung@google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>