CHROMIUM: arm64: kvm: Fix to set KVM vendor feature bits correctly
Fix to set the KVM vendor feature bits for new hypercalls correctly.
The ARM_SMCCC_VENDOR_HYP_KVM_FEATURES_FUNC_ID must return a bitmap but
the current code just set a number instead of feature bits corresponding
to the supported features.

This is a bugfix for CLs which are technical debts tracked by b:256983979.
Thus this will be also reverted when the original feature is upstreamed
correctly.

BUG=b:249685960
TEST=Build, deploy and run autologin.py to confirm arcvm boot.
UPSTREAM-TASK=b:256983979

Change-Id: If26fb75482d551b668948cc08df42967e9e094c4
Signed-off-by: Masami Hiramatsu <mhiramat@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4546812
Reviewed-by: Suleiman Souhlal <suleiman@chromium.org>