CHROMIUM: Add kdump configuration family
This configuration family will be used to build a kdump kernel. It will
be used to generate a core dump file of the kernel when it crashes.

Right now we only support brya. The current configuration is a clone of
the chromeos family. We will remove configurations later on, to have an
image as little as possible.

BUG=b:242048430
TEST=emerge-brya chromeos-kernel-dump-5_10

Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: Iaf68ceddb3c133a220a074c1c0207f52261afd25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4582597
Reviewed-by: Ross Zwisler <zwisler@google.com>