CHROMIUM: Fix build errors to access se.avg.util_guest with CONFIG_SMP=n
Since task_struct::se.avg is only available when CONFIG_SMP=y, the code
accessing that field must be enclosed by #ifdef CONFIG_SMP.

This is a bugfix for CLs which are technical debts tracked by b:256983979.
Thus this will be also reverted when the original feature is upstreamed
correctly.

BUG=b:249685960
TEST=Build with CONFIG_SMP=n
UPSTREAM-TASK=b:256983979

Change-Id: Ic3d5e30ead53e4e3632cd9b69c1a234f29595bbb
Signed-off-by: Masami Hiramatsu <mhiramat@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4555157
Reviewed-by: Aashish Sharma <shraash@google.com>