CHROMIUM: termina: Drop CONFIG_BOOTPARAM_HUNG_TASK_PANIC=y
Don't turn hung tasks into a kernel panic. We don't want this to be a
fatal condition.

When stopping a process that uses io_uring, the associated kworker
threads may enter the uninterruptible sleep ("D") state. This is
normally harmless, as execution will resume when the process is
continued, but they will be registered as hung tasks.[0] When this
option is enabled, a kernel panic occurs. Most notably, this can occur
when using neovim with a recent enough libuv.

[0]: https://github.com/axboe/liburing/issues/448

BUG=b:284728129
TEST=emerge-tatl chromeos-kernel-5_15

Change-Id: I922a70207519b1198ea8269f662156f765f922a3
Signed-off-by: James Ye <jamesye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4574247
Reviewed-by: David Munro <davidmunro@google.com>