CHROMIUM: Add CPPLINT.cfg to make "cros lint" happy
When using CrOS IDE and browsing a header file in the kernel the whole
file is covered with ugly warnings. This is because cros lint runs on
".h" files as if they were C++ files using the Chromium standard. For
reference, see:
* chromite code base: cli/cros/cros_lint.py (specifically the
  _EXT_TOOL_MAP)
* depot tools code base: cpplint.py

Let's add a few definitions to make "cros lint" happy.

This is, and will always be, a ChromeOS only thing. Add this as
CHROMIUM.

UPSTREAM-TASK=None
BUG=b:281517321
TEST=No longer see angry yellow warnings everywhere

Change-Id: Ia1ea4d726b3716572d15e9fc84cac0004bb3fd48
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4602468
Reviewed-by: Brian Norris <briannorris@chromium.org>