UPSTREAM: media: mediatek: vcodec: Fix potential crash in mtk_vcodec_dbgfs_remove()
The list iterator "dbgfs_inst" is always non-NULL.  This means that the
test for NULL inside the loop is unnecessary and it also means that the
test for NULL outside the loop will not work.  If we do not find the item
on the list with the correct the ctx_id then it will free invalid memory
leading to a crash.

Fixes: cd403a6a0419 ("media: mediatek: vcodec: Add a debugfs file to get different useful information")
Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>
Reviewed-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
(cherry picked from commit b19c347b123cd68d6208474194947bdb6dbc1227)

BUG=b:317227622
TEST=emerge-corsola chromeos-kernel-5_15

Change-Id: I605e059f1a2ea3eb288bd2f7f992bb920aa65983
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5197139
Reviewed-by: Yu-Che Cheng <giver@chromium.org>
Commit-Queue: Yu-Che Cheng <giver@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>