CHROMIUM: hrtimer: Add a sysctl to lower resolution of timers
On boot, by default the timers switch from periodic to high res mode.
However, if highres=false is passed, it switches to 'nohz' mode, which
is lower res.

This patch adds a sysctl to do the switch to 'nohz' mode at runtime, and
back to high res mode as well if requested.

[CHROMIUM note: I am marking it as CHROMIUM for experimentation in
Finch and will push upstream after seeing results.

Note that KVM depends on HIGH_RES_TIMERS in the kbuild so it is likely
we need this patch even in the instance we decide to disable highres all
the time and not dynamically. Also, we may want to exclude <1K HZ
devices such as ARM from that.]

BUG=b:263289152
UPSTREAM-TASK=b:289837863
TEST=cyclictest --smp -p95 -m with timer_highres=0

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>
Tested-by: Joel Fernandes <joelaf@google.com>
Change-Id: I43fa34c2d227569218178e78cfaef4fc3aec4013
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4660338
Reviewed-by: Vineeth Pillai <vineethrp@google.com>
Reviewed-by: Suleiman Souhlal <suleiman@chromium.org>
Commit-Queue: Joel Fernandes <joelaf@google.com>