CHROMIUM: add new cros_net tracepoints
!!!DO NOT UPSTREAM!!!
!!!DO NOT PORT TO OTHER KERNELS!!!

Add new tracepoints that will capture sk_buffers. This is needed to
determine which interface a packet is destined for so that internal
network flows can be ignored.

Also, in the future, the networking team's packet marking will need to
be examined. Those marking bits are stored within sk_buff.

The reason why this patch shouldn't be upstreamed is that the only
reason these tracepoints are added is to provide an attach point for BPF
programs. Normally this would be done by hooking onto fentry/fexit.
Unfortunately fentry/fexit is not supported in the 5.10, 5.15 kernel for
ARM platforms. Porting over the upstream patches to add this
functionality to 5.10 and 5.15 also isn't feasible. A port attempt was
made and the number of patches was large, many conflicts needed to be
resolved and the changes touched a lot of the bpf subsystem and some
changes extended outside of the bpf subsystem.

So rather than making wide, sweeping and probably breaking changes to
the kernel; a relatively few trace points were added. Once all ARM
platforms move to a kernel that supports fentry/fexit tracepoints this
patchset should be dropped.

Ideally this patch would never make it over to the new kernel , but if
it does we can always manually drop it.

BUG=b:279185537
TEST=tracepoints don't cause a kernel crash when turned on.
UPSTREAM-TASK=b:278303728

Change-Id: I8fe9fbfe8068df07e2453ae6e571b18f16c81366
Signed-off-by: Jason Ling <jasonling@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4704260
Reviewed-by: Guenter Roeck <groeck@chromium.org>