CHROMIUM: media: camx: drop CAM_CDM_BL_CMD_TYPE_KERNEL_IOVA handling
This appears to be dead code, remove and it just return -EINVAL
if we ever see CAM_CDM_BL_CMD_TYPE_KERNEL_IOVA cmd.

Per Qualcomm:
the code was added to accommodate the usage of passing kernel iova
in virtual cdm in the future, however it is unused and there are
no such plans as of now.

UPSTREAM-TASK=b:274966069
BUG=none
TEST=CTS and camera app on coachz

Suggested-by: Tomasz Figa <tfiga@chromium.org>

Change-Id: I57993029ec68eab532ca5e533f355c751eb4bbe4
Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4707446
Reviewed-by: Yunke Cao <yunkec@chromium.org>