CHROMIUM: Bluetooth: Add timeout in disconnect when power off
For some controllers, it is known that when the HCI disconnect and HCI
Reset are too close to each other, the LMP disconnect command might not
been sent out yet and the command will be dropped by the controller when
it is asked to reset itself. This could happen on powering off adapter.

One possible issue is that if we toggle the adapter power off and on
within a short time, the remotes we connected before powering off might
still think they are still connected to us, therefore, create connection
to those remotes will fail with ACL Already Connected error.

This CL makes the host stack to wait for |HCI_EV_DISCONN_COMPLETE| when
powering off with a fixed timeout (300ms). This value is set based on
experience that most disconnections take around 200 ms.

UPSTREAM-TASK=b:294496529
BUG=b:287557722
TEST=tested the below procedure with Logi K580 on a AX211 device.
1. pair the peer and stay connected.
2. power off and on immediately
3. observe that the btsnoop log is consistent with the configured
   timeout.

Signed-off-by: howardchung <howardchung@google.com>
Change-Id: I041f675579e0bfe1040ebb85ebc61d116ea7fea2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4707467
Tested-by: Yun-Hao Chung <howardchung@chromium.org>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Yun-Hao Chung <howardchung@chromium.org>