FIXUP: CHROMIUM: hrtimer: Add a sysctl to lower resolution of timers
Don't allow dynamic resolution change on devices lower than HZ=1000.
Doing so might severely effect performance and is uncharted territory.

Further, for experiments we rely on the availability of the
timer_highres proc node. Finch does not check for HZ, so we need the
kernel to not expose this node.

BUG=b:263289152
TEST=qemu
UPSTREAM-TASK=b:289837863

Change-Id: Ia128206906a09eec3e7604766c71858ecc1c285f
Signed-off-by: Joel Fernandes <joelaf@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4713005
Reviewed-by: Vineeth Pillai <vineethrp@google.com>