CHROMIUM: config: Switch CONFIG_TYPEC from =m to =y for Qualcomm
Matching what we did in the fallback config in
<https://crrev.com/c/4717925>, let's switch CONFIG_TYPEC to =y for
Qualcomm. While this isn't truly needed in the current chromeos-5.15
kernel, if we end up getting the patch from a backport we could run
into the same issue we saw with upstream. Having this =y isn't a huge
deal.

NOTE: the old value of `CONFIG_TYPEC=m` was forcing
`CONFIG_CROS_EC_TYPEC=m` because the cros_ec symbol depended on the
typec symbol and a builtin can't depend on a module. Now that we've
made the typec symbol a module we have to be explicit about wanting
`CONFIG_CROS_EC_TYPEC=m`.

BUG=b:293151872
TEST=Display comes up

Change-Id: If276f42657ac560302d06a103f13414b2174ed68
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4717970
Reviewed-by: Stephen Boyd <swboyd@chromium.org>