CHROMIUM: venus: hfi: Fix bounds checks for filling capabilities
Fixes "FROMLIST: venus: hfi: add checks to handle capabilities from
firmware". When HFI version is 1XX, the core capabilities will not be
memset to 0. The existing |num_pl|, |num_caps|, and |num_fmts| value may
be greater than zero. The updated checks validate that the last memcpy'd
bytes will not write past the end of the arrays that they are filling.

UPSTREAM-TASK=b:295178805
BUG=b:286974958
TEST=Compiles. Will watch Trogdor test results

Change-Id: I847fc1b0463f66156ea6f47e6344759b924bcbcc
Signed-off-by: Nathan Hebert <nhebert@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4766225
Reviewed-by: Stephen Boyd <swboyd@chromium.org>