CHROMIUM: msm: camera: mem: Add check for number of sof event data being stored
In __cam_req_mgr_setup_link_info, we are not checking the value of
link->num_sof_src, which is used as an index for array
link->dev_sof_evt[]. This can cause an OOB write issue. This change adds
a check for this condition to prevent OOB write.

BUG=b:295172896
UPSTREAM-TASK=b:295172896
TEST=Do tests in the prerequisite CL

Change-Id: I0d27d616a11f5da10e165a6cb5241406d69744a3
Signed-off-by: Shardul Bankar <quic_sharbank@quicinc.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4780059
Commit-Queue: Sergey Senozhatsky <senozhatsky@chromium.org>
Reviewed-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Tested-by: Sergey Senozhatsky <senozhatsky@chromium.org>