sched: Relax sched class linkage check
When GCC LTO is enabled, the sched classes are indeed emitted in the
correct order but may not necessarily be adjacent to one another in memory.
As a result, the linkage check in sched_init() is spuriously triggered with
GCC LTO.

Relax the check by strictly checking for the order of each sched class as
intended.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
(cherry picked from commit ca448ef8f2fb8b12a90ded75fe432979e416b277)