CHROMIUM: drm/i915/vrr: Allow pipe mismatch in VRR params for DRR fastset
In case of a panel that supports DRRS and seamless_m_n, dual refresh rate
allows to seamlessly switch to a downclock mode to save power.
This is expected to happen seamlessly without a full modeset as in this
case we allow the mismatch in mode's crtc clock.

With VRR fastset patch series (https://patchwork.freedesktop.org/series/115422/)
it allows VRR to be toggled in fastset by pre computing and pre programming
VRR parameters irrespective of VRR enabled CRTC state. Now in case
of Dual refresh rate (DRR), when we switch to/from downclock mode, crtc_clock
changes and these parameters get computed to new values.

This patch allows the VRR parameters mismatches in case of seamless_m_n fastset.
This will allow Dual refresh rate to throttle seamlessly w/o full modeset.
These VRR parameters will need to be reprogrammed to the HW in fastset for
DRR + VRR to operate in fastset fashion together.
This will be handled in follow up patches that are being reviewed
upstream. So until we handle this, disable VRR in kernel whenever
it is in DRR downclock mode.

BUG=b:279750772
UPSTREAM-TASK=b:297940747

Change-Id: I3f97b6024051caa1e1d714faf57b82b65ae197a7
Signed-off-by: Manasi Navare <navaremanasi@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4823704
Reviewed-by: Sean Paul <seanpaul@chromium.org>