BACKPORT: FROMGIT: arm64: smp: Add arch support for backtrace using pseudo-NMI
Enable arch_trigger_cpumask_backtrace() support on arm64. This enables
things much like they are enabled on arm32 (including some of the
funky logic around NR_IPI, nr_ipi, and MAX_IPI) but with the
difference that, unlike arm32, we'll try to enable the backtrace to
use pseudo-NMI.

NOTE: this patch is a squash of the little bit of code adding the
ability to mark an IPI to try to use pseudo-NMI plus the little bit of
code to hook things up for kgdb. This approach was decided upon in the
discussion of v9 [1].

This patch depends on commit 8d539b84f1e3 ("nmi_backtrace: allow
excluding an arbitrary CPU") since that commit changed the prototype
of arch_trigger_cpumask_backtrace(), which this patch implements.

[1] https://lore.kernel.org/r/ZORY51mF4alI41G1@FVFF77S0Q05N

Co-developed-by: Sumit Garg <sumit.garg@linaro.org>
Signed-off-by: Sumit Garg <sumit.garg@linaro.org>
Co-developed-by: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Misono Tomohiro <misono.tomohiro@fujitsu.com>
Tested-by: Chen-Yu Tsai <wenst@chromium.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Link: https://lore.kernel.org/r/20230906090246.v13.4.Ie6c132b96ebbbcddbf6954b9469ed40a6960343c@changeid
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
(cherry picked from commit 331a1b3a836c0f38165dcec168c0a03b93cf0c17
 git://git.kernel.org/pub/scm/linux/kernel/git/arm64/linux for-next/backtrace-ipi)

Conflicts:
  arch/arm64/include/asm/irq.h
  arch/arm64/kernel/smp.c
...technially they didn't cause conflict with the "git cherry-pick",
but those files needed to be manually modified because we don't have
commit 8d539b84f1e3 ("nmi_backtrace: allow excluding an arbitrary
CPU"). Picking that gets into picking the whole pile of cleanup
patches relating to the upstream of the buddy hardlockup
detector. That's a lot of work for the 5.15 kernel when the downstream
detector works fine there. This backport is trivial (just change
"exclude_cpu" to "exclude_self").

BUG=b:197061987
TEST=echo HARDLOCKUP > /sys/kernel/debug/provoke-crash/DIRECT

Cq-Depend: chromium:4919746
Change-Id: Ie6c132b96ebbbcddbf6954b9469ed40a6960343c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4704887
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>